Extract tuple
