+++ src/main/kotlin/practice/TuplePractice.kt
@@ -1,0 +1,4 @@
+package practice
+
+class TuplePractice {
+}
