+++ src/test/kotlin/practice/TuplePracticeTest.kt
@@ -1,0 +1,5 @@
+package practice
+
+internal class TuplePracticeTest {
+
+}
